requirements.txt
absl-py==2.3.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
Cartopy==0.24.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dask==2025.5.1
flatbuffers==25.2.10
fonttools==4.58.1
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.71.0
h5py==3.13.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras-nightly==3.10.0.dev2025060104
kiwisolver==1.4.8
libclang==18.1.1
locket==1.0.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-venn==1.1.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
narwhals==1.41.0
numpy==2.1.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.0
Pygments==2.19.1
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
simsimd==6.4.7
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
stringzilla==3.12.5
tb-nightly==2.19.0a20250218
tenacity==9.1.2
tensorboard-data-server==0.7.2
tensorflow-hub==0.16.1
termcolor==3.1.0
tf-keras==2.15.0
tf_nightly==2.20.0.dev20250530
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
